+++ examples/resources/aiven_flink_jar_application_deployment/resource.tf
@@ -1,33 +1,23 @@
-resource "aiven_flink" "example" {
-  project                 = data.aiven_project.example.project
-  service_name            = "example-flink-service"
-  cloud_name              = "google-europe-west1"
-  plan                    = "business-4"
-  maintenance_window_dow  = "monday"
-  maintenance_window_time = "04:00:00"
-
-  flink_user_config {
-    // Enables upload and deployment of Custom JARs
-    custom_code = true
-  }
-}
-
-resource "aiven_flink_jar_application" "example" {
-  project      = aiven_flink.example.project
-  service_name = aiven_flink.example.service_name
-  name         = "example-app-jar"
-}
-
-resource "aiven_flink_jar_application_version" "example" {
-  project        = aiven_flink.example.project
-  service_name   = aiven_flink.example.service_name
-  application_id = aiven_flink_jar_application.example.application_id
-  source         = "./example.jar"
-}
-
 resource "aiven_flink_jar_application_deployment" "example" {
-  project        = aiven_flink.example.project
-  service_name   = aiven_flink.example.service_name
-  application_id = aiven_flink_jar_application.example.application_id
-  version_id     = aiven_flink_jar_application_version.example.application_version_id
+  project        = "my-project" // Force new
+  service_name   = "my-application" // Force new
+  application_id = "foo" // Force new
+  version_id     = "543e420d-aa63-43e8-b8e8-294a78c600e7" // Force new
+
+  // OPTIONAL FIELDS
+  entry_class        = "com.example.MyFlinkJob" // Force new
+  parallelism        = 1 // Force new
+  program_args       = ["example-argument"] // Force new
+  restart_enabled    = true // Force new
+  starting_savepoint = "path/to/savepoint" // Force new
+
+  /* COMPUTED FIELDS
+  deployment_id  = "1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d"
+  job_id         = "foo"
+  created_at     = "2021-01-01T00:00:00Z"
+  created_by     = "foo"
+  error_msg      = "foo"
+  last_savepoint = "foo"
+  status         = "CANCELED"
+  */
 }